requirements.txt
openai
networkx
langchain-text-splitters
python-dotenv
loguru
